Vastly improve the default plugin selection heuristics.
This now takes into account DLC, ESM-flagged masters, ITPO records
(Identical To Previous Override), and which face edits contribute to
facegen data. The UI also makes a distinction between any face edits
(old winky face) and risky edits (black face).

With all of these updates, it is now able to correctly pull the set of
default records entirely from the "core" plugins (USSEP, CRF, DLCs) in a
minimal load order with a few dozen added NPC mods, while pulling faces
from the new mods.

There are still a lot of warnings from USSEP and CRF that have to be
suppressed, but there seem to be explanations for most if not all of
them, e.g. records/facegens that were partially corrected in DLC.