Convert XEditLib code to interfaces and generics.
This is in preparation for the planned switch to Mutagen, and should support a breakage-free transition.

Due to inherent limitations of C# generics, we have to reify all the page classes and can't reify anything under them, so each page has a using directive to fake some consistency. It means that when switching over to Mutagen, which uses different record key types, all of these will have to be updated at once, but at  least it's only a half-dozen or so.