Switch Mugshot to be a class type.
This fixes a bug in the Profile Page that was "fixed" once before by changing the way double clicks were handled, but turned out to still be a problem. It appears to be due to a similar scenario as https://docs.microsoft.com/en-us/troubleshoot/dotnet/framework/argumentexception-select-row-wpf-datagrid.

Specifically, even though we weren't using a custom Equals/GetHashCode implementation, by using a record type, C# was doing it for us, and causing the "key already added" exception due to the use of mutable properties, which aren't forbidden or even warned against for use in C# records. Seems like they're "mostly" fine, except in certain scenarios where they cause subtle problems.