Destroy Loader control when loading is done.
Simply being hidden means it continues to monitor the LoaderViewModel,
and in particular continue monitoring and displaying log entries, even
when no log UI is really "active". With detailed logging in place (in
progress), this causes visible slowdowns to the tasks themselves.

Destroying is achieved by switching to a ContentControl and
ContentTemplate. There are cleaner ways to do this than style triggers,
but they don't work properly with generic types.