Add "basic" profile autosave/autoload.
Implementation writes log entries on every relevant NPC change, and restores on startup. Profile event log is stored in the same local directory as regular execution logs. It will also detect new NPCs and fields and ensure they get written on the first sighting, so they'll remain stable even if the load order is changed without the user having made any explicit changes to their values.

It *should not* log any "implicit" changes due to the app having been loaded with previously-enabled mods or plugins missing, *unless* the user explicitly corrects those errors, i.e. to complete a build. In other words, the autosave will survive the program being accidentally started with the "wrong" MO2 profile and then restarted with the "right" profile.

However, there is no log trimming, so this isn't designed for intentional large-scale profile changes, e.g. repeatedly opening totally different configurations. Each instance of doing this will add all of its related changes to the log and eventually grow the log to an unmanageable size. Since this is alpha, and those kinds of changes are not the typical expected use case, this "basic" implementation should suffice for a while.