Store Profile focus/selections in view model.
This allows most of the UI state to be preserved even after switching to a different page and back. Some code behind has been switched to bindings and property change handlers, which in turn requires being a bit more selective in when to "clear" previous state.

Also adds a short code-behind line to scroll the NPC grid to the selected item, which doesn't seem to be achievable through the view model.

Filter states are already preserved thanks to our quirky header binding implementation, but preserving sorting (which uses the default DataGrid behavior) looks to be a huge PITA and will have to be done as a separate future improvement. Hopefully, losing sort state won't be as painful as long as the grid automatically scrolls to the selected item.