De-wiggification v1.
Works by attempting to find a matching hair in a resource mod like KS Hairdos. If found, adds those hairs to the NPC's head parts and also adds/replaces hairs in the FaceGen NIF. With a fair amount of only random spot-testing, this is reliable in game and does not cause blackface.

The only mod that has been tested with it is High Poly NPC Overhaul 2.0, as I'm not aware of any other mods that employ this kind of strategy. And it's a kludge - some hairs won't match properly due to misspellings by the author ("SweetVillian"). Nevertheless it does essentially work for the hundreds of NPCs where the heuristic is effective, and if users don't ignore the baldness warnings for the broken ones and choose other mods for those (or just revert to vanilla), then it will be fine.