Refactor: Extract Main namespace.
Also does Maintenance which is tiny (for now) but still its own page.

To avoid having circular dependencies on Main, due to the need to use the MainViewModel as the DataContext for individual Pages, the MainViewModel now implements a bunch of one-line "Container" interfaces that _don't_ depend on Main and that these other pages can use instead.

XAML still references MainViewModel, but that's really just for design-time support and shouldn't be a real problem.