Move Nifly code to its own namespace.
While this is a trivial amount of code, it still brings in an external dependencies and should therefore be isolated from other external dependencies.

This final refactoring of the root makes the root solely for the app entry point, general build/VS support, and truly "common" UI components that could be used on any page.