Allow disabling de-wiggification before build.
This affects both the build and the pre-build checks, so it's listed as an option in the pre-build panel and affects the IDs and text of warnings that might be generated. To ensure it gets displayed or hidden correctly based on the most recent selections, this requires a slight "refresh" when the Build page is activated from the nav bar.