Fiddle with assembly name a bit.
With libraries we generally want the full namespaced name, but with executables this is highly visible (and kind of annoying) in e.g. Windows task manager, so it really should just be the application name.

Also updates the hardcoded name in log entries, which should be the last hard reference to the old name.