Upgrade to latest Mutagen.
This is partly to incorporate a few bug fixes (such as those relating to the archive list) but primarily to use the newer GameFinder, because Steam has changed their implementation and recent installs (after June 6) can no longer find the game.

Tidies up a few areas of the assembly info and versions. Unfortunately this causes .NET to use a new settings location, which can't be helped without a custom settings provider.