Overhaul the overhaul detection.
Motivation originates from a report that AI Overhaul (the newer one, not Immersive NPCs) was frequently not being detected as the default plugin. The somewhat naive decision of picking "last mod without face edits" was pretty good for avoiding face mods as the default, but also nerfs a lot of behavior mods that happen to have face edits.

These changes look explicitly for behavior edits, and ambiguous edits that might be important in a non-overhaul mod (outfit, body), while also doing a second pass of "overhaul mod detection" to strip away the last remnants of accidental and opinionated edits. Like most algorithms here, it is unlikely to perfect, but it does appear to be handling the combination of USSEP, CRF, and AI Overhaul very well as defaults, along with any combination of finicky NPC mods as faces. It made zero mistakes in a few manual tests.

This also includes an update to copy height, weight, and animation flags from the overhaul mod, especially since they are now being considered in detection of behavior mods. Even though these aren't technically face attributes, in most cases it is going to make more sense to follow what the visual overhaul wants, rather than some other arbitrary source.