Handle ITMs in standard rule set.
Previous implementation was ported from the NpcConfiguration, but the list of overrides in NpcConfiguration includes a fake override representing the master. This case wasn't handle in the rule set, causing it to infinite-loop when the ITPO plugin name was also the master, and therefore did not correspond to any override in the list.

Fixes #15.