Improve usability and safety of the load order screen.
The original version was very basic and did not do any checks whatsoever, leaving it possible to choose a load order with missing masters or old patch output, which would appear to be OK during loading but crash the app immediately afterward.

The updated version identifies plugins that can't be loaded, e.g. due to masters that are missing or simply disabled, and provides a prominent visual indicator that some plugins won't be loaded. Additionally, if some plugins are unloadable, they will simply be ignored, so the app may load with fewer plugins than the user expected but won't crash.