Add mugshot redirects.
Includes Settings model and UI for changing it, as well as the logic to handle these redirects in the `ProfileViewModel`.

Redirection rules for `GetMugshots` is actually a little confusing, because that method obtains the mugshot _first_ and tries to work out the mod afterward, meaning that there may be multiple matches. The rules are designed to try to prioritize whichever will give the best/safest result.

Fixes #3.