Adds all planned profile filters.
As in the feature description, this includes all of the various conflict and "danger" flags, as well as searching by specific plugin selections. This should make it much easier to troubleshoot problems with a profile.

Filters don't immediately update until after editing an NPC, unless they are unapplied and reapplied - this is on purpose, as otherwise it could become too easy to make a mistake and immediately get lost and be unable to find the record again.

One limitation is that even though the filters work, it is difficult to understand exactly what the problem is when an NPC references an invalid mod and the plugin is no longer available - it will simply show up as having a different plugin, selected by default, which is not ideal. The UI could make it clearer that the plugin is supposed to match, but the correct plugin is missing. However, this is more about the plugin selector and not the filters, so it belongs in a new issue.

Closes #5.