Add filtered (invalid only) NPC reset.
Does exactly what's described in the issue and limits the reset to NPCs that are pointing to unreachable plugins, if the corresponding option is checked.

Also adjusts the build checker to only pay attention to NPCs that are actually in the current profile, otherwise it's inconsistent with the reset and with the build itself. Sometimes NPCs can be filtered out of the profile, even if they have invalid selections, due to no longer having any overrides. These are ignored in the output, so there is no reason to report errors on them.

Fixes #18.