Jump to profile row from build warning.
The implementation of this is a hack, since pages can't really talk to each other, and especially since the row isn't guaranteed to be visible through the filters. The hack forces it to be visible until the filters are changed, which actually requires multiple property change notifications and cycles to work correctly, but doesn't seem to be too noticeable in the UI, performance-wise.

Only really noticeable effect is that the row will sometimes appear at the top and sometimes in the middle, but is still always within the visible area. Should be an acceptable tradeoff.

Jump is activated by double-click. For now there's no additional link in the info panel, but it could be added easily if there's a need.

Introduction of `RecordKey` wasn't strictly necessary here, but it's a first step toward refactoring something that's very repetitive.

Fixes #14.