Use resolver for directories <--> mods.
This allows us to have a many-to-one mapping from directories (which are "files" in Nexus/Vortex world, #9) to mods, while still doing all of our selections by mod.

One limitation is that it is subject to whatever order is in the Vortex manifest, which might not reflect proper priorities in terms of patch files - it always just takes the last one available if there are multiple copies. If this causes problems, it may be required to either look at the "primary/main" bit in the mod metadata, or try to hook into the actual Vortex conflict-resolution settings. But for now this works pretty well.