Warn on Vortex start without manifest.
This uses two checks: (1) the parent process, so as to detect first-time launches before a mod directory has been configured, and (2) the presence of the staging marker ("__vortex_staging_folder") in the mod root directory, for subsequent launches even if the tool has been launched from _outside_ Vortex.

If Vortex is detected without the accompanying manifest argument, it will display a warning pointing users to the Vortex extension hosted on Nexus, and - we hope - avoid a lot of negative experiences and spurious bug reports from users who might not be aware of the special requirements for Vortex. This is important for actually being able to claim to support Vortex (#9).