Fix duplication of default redirects.
The expected behavior here is to use either the entire default list (e.g. of redirects), if the saved settings doesn't contain its own list, or the entire saved list if it does. Using the `Populate` method was actually merging both lists, creating a duplicate each time the app was launched and potentially multiple duplicates if the changes were saved.

The easy workaround - clearing the lists before `Populate()` - doesn't work because it then won't use the default list at all, even if the saved settings has no entry. Instead, we have to deserialize to a new object and do field-by-field copying.