Import mod directory from Vortex.
Vortex extension now adds the mod staging directory to the manifest file. When starting EasyNPC for the start time, it will use this to set a default value for the mod root directory. The intro screen still shows, but no additional steps are required other than clicking "I'm done".

This intentionally won't overwrite the staging directory if it happens to be empty but it is NOT the first launch - or if first-launch mode is forced via the command line but the existing setting is non-empty. It will only trigger when we can be 100% certain that it is not overriding the user's intent.

Closes #9.