Add preliminary Mod Organizer integration.
MO2 "deep" support relies on the assumption that the main MO2 INI is in the same directory as the executable - so if EasyNPC is being launched from MO2 then there is no need to explicitly ask for this info, it can be detected via the parent process info.

Currently there is nothing special to do for mod-file mapping, since the assumptions were designed around MO2 in the first place. What it does do is detect the Mod Directory by default, using the INIs, so it actually uses the configured directory rather than making naive assumptions, and potentially saves a lot of clicks and confusion on the first run.

Detection of startup environment is now also less hacky - specifically doing detection of mod managers rather than slapping together a lot of arbitrary conditionals.

Some of the first-time messaging has also been improvement. Text descriptions of documentation have been replaced with hyperlinks, a new paragraph was added for when defaults were detected, and the "I'm Done" button now more clearly indicates if there are still problems to fix, although it will still allow ignoring those problems.

Improves, but is not a full implementation of, feature #10.