Make created mod visible in Vortex.
This uses a hacky file-based "IPC" to inform Vortex of when and where the new mod is, and make Vortex recognize it immediately as an installed mod.

Prior to this, the mod would only be visible after restarting vortex and confirming a warning message. Now there is no need to restart, and no warning after a restart.

Ref #9