Extract file/archive code into new libraries.
This is a small step toward improving the service definitions for DI (#26) but is more immediately relevant in the context of building a new standalone tool to fix up facegen files in preparation for mugshot creation. The main function these two have in common is the "vfs" composed of the game's loose files and loaded archives.