Fix a few Mutagen quirks.
Somehow this was glossed over in EasyNPC development because of the way the GameEnvironmentState is manually constructed, but the "default" behavior of Mutagen is to set the "data" directory to the game directory (without "data" in the path).

This adds (and uses, where applicable) an extension to resolve the discrepancy, if it exists. Also tweaks the Archive Provider to follow the load order for BSAs, which isn't normally important except when using file providers.

Includes required GPLv3 license due to the dependency.