Support named Mod Organizer instances.
The fix for #29 resolved the major problem with the app not starting, but did not actually include any instance detection, so the mod directory would be blank by default for those installations.

Mod Organizer stores the current instance in the registry, and instances themselves in LocalAppData by the same name. Using this, we can get the _correct_ INI file for the current Mod Organizer instance.