Ignore Vortex mod entries with no name.
For reasons not currently known, some of the Vortex mod metadata entries are empty - they don't include a mod name. These result in the `ModPluginMap` trying to add null keys to a dictionary, which results in the associated bug.

As a workaround, we can do an explicit check for missing mod names.

Fixes #34.