Add null checks for face tints.
These values shouldn't ever be null, and weren't null in a repro attempt, but apparently are null in some instances. It's not going to fix potential problems with equality detection, but we can prevent load errors by ignoring tints with values that really _can't_ be null (like the index/color) and substituting defaults for the ones that are borderline.

Without knowing which field is causing the problem, have to guard all of them.

Fixes #47 (probably).