Ignore unresolvable races for head parts.
Some plugins are putting strange things in there, like pointers to keywords rather than races. The try/ignore pattern will also ignore actual invalid races records, which isn't great - but in some cases there won't be any way to differentiate between an invalid race and an invalid not-race, so this may be the best we can do. Either fail on startup where 99% of the time it doesn't matter, or fail/create unexpected results in the build the other 1% of the time. 99% is more important right now.

Fixes #36.