Improve safety of archive handling.
Instead of letting exceptions bubble up and cause crashes, handle them and ensure they're logged. If an exception indicates that an archive is corrupted at the header (i.e. totally unreadable), don't try to use it again.

Also logs which archives are being indexed to provide some additional debug info in case an exception is still able to escape.

Fixes #63.