Improve messaging for missing games.
Running on unsupported games is now the _most_ common reason why the app fails to start, and in the majority of these instances the problem is not correctable (the game just isn't supported) and crash reports don't help.

Make it clear to the user what's going on and why.

This will still allow crash reports if GameFinder itself _crashes_. The new behavior only activates if GameFinder succeeded but couldn't find the game. In these cases, there is no way to get any useful information from the stack trace.

Closes #62. Also is written fairly generally to keep a path open for #32.