Discourage use of Trim operation.
This is a necessary feature but applies in a very, very small number of scenarios, and anyone using it to try to solve a general problem with build warnings (as opposed to sluggish startup/UI performance) is going to end up causing more problems.

Add a prominent warning and change the button to danger color.

Fixes #40.