Add quick and dirty compatibility system.
Rather than just write a few lines in the MutagenAdapter to ignore children completely and maybe log something, this is anticipating other such issues to come up in the future, not just in this app but in future modding apps, and trying to come up with a way to organize and report them effectively.

It's far from a perfect implementation, given the time constraints, but will allow us to organize all the rules in one place and report the rules in use when starting the app. Logging is built into the rules themselves, so that whenever something fails a check, it is automatically reported.

Ideally, overrides could also be individually suppressed this way, but the UI needs to be done properly, i.e. actually showing these overrides, but with warnings and not allowing them to be used. Simply eliminating overrides would make things worse by causing reports of missing plugins in existing profiles and adding general confusion when records don't show up. So for now, this only operates on the "main" record.

Fixes #60, temporarily anyway, until these mods are supported properly.