Detect "implicit vanilla" face mods in profile.
When restoring the autosaved profile, mods that are pointing to vanilla for the face plugin (i.e. don't have any other facegens, or have actually been overridden to stay with vanilla) will have no autosave record at all for the face. Due to the current design of the autosave and restore, this could previously leave the NPC config pointing to a face mod that is incompatible with the restored plugin, e.g. one that was newly found and _did_ match the face mod until it was overwritten by the autosave.

A more concise way of saying this is "don't allow NPCs already in the profile to have their face mods reset on startup", which is the intended effect but not the precise mechanism.

Fixes #65.