Allow setting null face mod on reset.
When resetting face mods for missing refs, it _was_ finding the correct plugin to reset to, but there were two problems.

First, since the plugin had already been "implicitly reset", it wasn't being emitted as a change event since the value appeared to be the "same. This has been updated to keep track of the old, invalid name on profile restore, until it is changed again to a known override.

Second, the face setter explicitly avoids clearing the face mod if it can't find a match. This makes sense during normal operation - there's no reason we should keep forcing the face mod back to vanilla just because the app is having a hard time matching the plugin. But on an explicit reset attempt, we have to override this and allow it to revert to vanilla.

Another way of doing this would be to detect a vanilla plugin name and use that to clear the face mod. If the above solution turns out to be unsafe for some reason, then that's a possibility. But it covers more potentially unknown and weird scenarios to use the current logic - that way, after a reset, the mod pointer will always be _valid_ even if it's wrong.

Fixes #66.