Use message bus for forced navigation.
Message bus will be a general concept for components to communicate in a decentralized way without having to plumb event handlers and such through multiple layers. Page navigation already has to do this, so it has been converted, and a similar method will be needed for navigating to a filtered profile view (#17).