Highlight/animate filter button on change/force.
This should make it much more obvious to users how to "recover" from double-clicking on a master dependency in the build screen.

Current implementation checks for default filters, not empty filters, so the button doesn't light whenever something is hidden (i.e. DLC-only changes), but when filters are changed.

Fixes #17.