Reset build state on certain changes.
This is a hack to deal with a flow that has gotten a little weird. If the user has previously dismissed warnings (i.e. they're at "ready to build") *and* has subsequently changed the profile, then the build state will reset so that checking is recommended again.

Especially important for dealing with unwanted masters, because otherwise they'd come back to the build screen and it would look like the master is still required, even if it no longer is.

Fixes #17 (more).