Exclude NPCs without facegen.
There are a lot of these, and they pollute the profile page and output mod for no gain, since their faces can't be customized.

This effectively means "humans only" but there are a few exceptions, like DremoraRace.

Because the rule is hit so often, log entries can be spammy, so a log level has been added to the rule system and this new one only outputs at debug level.

Fixes #70.