Add face tint fixing to MugPrep tool.
Refactors so that cubemaps and facetints use the same code path, since it's almost identical but for different nodes.

Also updates the cubemap fix to recognize nodes without eye data - some facegens don't have this, so we need to use the head part data in the plugin records to determine which node types are which head parts. (And we have to do this anyway for face nodes)