Start of analysis project / WIP.
This is meant to replace EasyNPC's own NPC analysis and start augmenting it with several other checks and types of analysis, and also be used for other projects down the road.

The goal is to have a modular system where different analyzers can be set up at run on a load order, to produce various kinds of descriptions of records (and, later, of plugins themselves). The first candidate is obviously the NPC, which has several existing checks to run already.

Since these haven't been fully tested and can't be implemented as a literal copy/paste, they're not integrated into the main app yet, just a standalone project.

Ref: #26