Add scrolling to the pre-build master list.
Don't want the build button to always be at the bottom, and putting a scroll viewer on the entire group is buggy because the listbox tries to capture the wheel, and making the listbox a fixed size would look weird when maximized. So, this takes on a new package to put a math expression on the list box's max height.

Fixes #77.