Accept game arguments from command line.
Includes the name and an override path to the data folder, should the automatic detection fail. Plumbs it through all the Mutagen abstractions so that no references are hard-coded anymore, except in the low-support standalone tools.

Adds a separate error message on startup for when the game ID is invalid, and renames the exception/templates so they're not ambiguous in this regard.

Progresses #32. Needs testing before anything can be closed.