Add first unit tests.
There's a mountain ahead, but we have to start somewhere, and making sure that our `RecordKey` is safe to be compared or used in sets and dictionaries anywhere seems as good a place as any to start.

As part of this, changes that type so that _all_ comparisons, including implicit ones, are safe in this way (ignore case), without having to opt in using a comparer. There's no reason we would ever want to do case-insensitive comparisons on this type.

#27