Implement common tests for NPC Analyzer.
This makes use of an abstract base class for analyzer tests that will verify the "basic analysis" fields automatically - now used for both the basic and NPC tests.

Because Moq was misbehaving (again), several things had to be rewritten, and a few bugs in the tests themselves were fixed, i.e. basic tests that didn't actually set up the records they needed to verify.

#27