Add DI helpers for the AnalysisRunner.
The AnalysisRunner is fluent, but requires a lot of external dependencies (e.g. on Mutagen) to set up. To allow us to abstract the whole thing, put it behind an `ILoadOrderAnalyzer` that only depends on the load order and a few simple params.

Includes a few tweaks to the `AnalysisRunner` after testing - adds a check for elapsed time (helps with performance monitoring) and moves `AsParallel` below the `Where` so that ordering is "sort of" consistent, i.e. it actually processes the largest (vanilla/mega-patch) plugins first. And now allows optional processing of implicits like Skyrim.esm.

#26