Adds Record Analysis chain model.
Load order analysis is structured top-down, but this isn't the view that's important to EasyNPC - or other tools and patchers, most likely. Instead, we want a record-oriented view - a list of all the "record chains" (master + overrides) of a particular type.

This can be computed entirely based on the prior analysis, and is much easier to consume in view models and other application code.

#26 #27