Add NPC name to analyzer.
Although this isn't really an "analysis", if we want to base the new NPC view models on this existing analysis type without requiring a separate pointer to the NPC "data", then this needs to be part of the analysis result.

We already provide the Editor ID, which is a similar concept, so the design isn't really changing.

#26 #27