Better memory management for new analyzer.
First implementation was somewhat naive - as long as a default analyzer was configured, it would process all records. All of these records and (for Mutagen case) their containing groups get permanently cached. For many record types, this doesn't add a ton of bloat, but it was massive for Worldspace records and fairly large for a few others.

Record types can now be explicitly ignored. In addition, the Mutagen cache can be (and by default is) purged at the end of analysis, since it's likely not going to be needed for a long time, if ever, once the analysis is actually produced.