Add services for file indexing.
Includes abstractions and a filesystem-based implementation for indexing (i.e. for very fast lookups) all of the files in a particular location. This is going to be used in a variety of places including the new mugshot implementation, mod-manager support, and indirect dependency-aware builds.

Archive indexing will likely use the same interface (`IFileIndex`) but will be a separate implementation because archives are read-only - don't need the watcher or bucketing, except if the latter is used to bucket e.g. by archive name.

#10 #26 #27 and probably more.