Enforce package order in NPC analyzer.
For most form lists, we do actually want a "set equals" (renamed to be consistent with that behavior). For packages, however, the order is important, and USSEP and probably other mods sometimes do nothing _but_ change the package order for a particular NPC. This is actually a change in behavior.

#85