Implement an indexed mod repository.
Accepts an existing file index (so e.g. it can be completely destroyed and recreated if the target directory changes) and automatically manages archives and tracks additions/deletions to the file system.

Component resolution is handled by a dependency, so that this should be able to work for any mod manager. For Vortex, this comes from the bootstrap file (metadata "DB"), and for MO2, it is read from various INI files.

#10 #22 #26 #27