Handle duplicate mod names in index.
And by "handle", this means "toss". This came up with a mod - Apachii Divine Elegance Store - where the main part of it was for SSE, but there was another mod installed (UUNP bodyslides) for Oldrim, which had the same name.

Only logical thing to do here seems to be to index both mods, but pick a winner (the first one) when searching explicitly by mod name. The "dupe" can still be found using any other search method.

#10