Parallelize initial mod component resolves.
Gives us a modest improvement in worst-case startup time.

Best-case scenario is to do the indexing at the same time as load order selection/record analysis is happening, and have it finish before any user-visible wait. This still helps for that scenario because it makes the indexing more likely to finish by that time.

#10