Read enabled mods from MO2 profile.
Looks at the selected profile (if known) and reads the state from the mod list file (if found). This makes the app aware of which MO2 mods (AKA mod components) are actually enabled and should help a great deal with mugshot display and builds themselves.

This effectively completes the Mod Organizer integration feature, everything that really matters at the moment is included. The rest comes down to UI treatment, i.e. altering some warning text, changing which resources are searched/copied in the build, etc.

Closes #10