First batch of new profile code.
This is the minimum required to display and interact with a profile based on the new analysis and mod repository code, not counting the XAML changes to the main Profile page. Most of the WPF components are just extracted from the original Profile page, but made standalone in order to provide better view/viewmodel encapsulation.

Not "fully" tested but all of the basic flows seem to work when integrated with the main app (TBD in a later commit).

#85 #64 #61 #51 #49 #37 #26 #22 (doesn't complete any of these, but advances all of them)