Extract Mods Directory into common interface.
Several different parts of the stack (e.g. app startup and settings UI) need access to the "default mod directory" concept we have on the old `IModRepository`, but previously it was only available on the MO2-specific configuration interface. Move it to a higher-level interface that's common to all mod managers.

#26