Improve Mutagen abstractions.
Should be the last of what's necessary, aside from Mutagen code still in EasyNPC itself, to get Mutagen working with IoC containers and the app startup flow.

Aside from just abstracting a few more things, like parsing/detecting the game installation, this adds a "game setup" interface that deals with all the information that can be known _prior_ to loading the game, and does not require any user input or confirmation to obtain.

#26 #27