Add late-configurable mod repositories.
Because the repository behavior is user-controlled - i.e. can choose the mod directory - this needs to have a configuration method in order to avoid doing a bunch of async/busy-waiting work in the constructor, having to wait a long time to actually construct it, and generally getting into dependency hell with any IoC container.

With this version, the mod repository can be configured and constructed right away by an IoC container, and then told to start indexing after e.g. confirming that the mod directory is configured and exists, or after the welcome screen or splash screen.

#26 #27