Rewrite and restructure build checks.
Each build check is now independent, and implements an interface so that an IoC container can inject them all into a build checker as a list/enumerable.

Checks have also been rewritten to use the new analysis system, and not have any generic parameters. The facegen check in particular now compares head parts and produces virtually no false positives - the only ones remaining are template NPCs, to be handled in #90.

A few minor features may be missing and labeled as todos.

#26 #37 #50 #64