Add mugshot-related observables.
Since the Mugshot directory can be changed while the app is running, we need a way to conveniently and automatically rebuild the mugshot index when that setting is changed, while still being able to inject a single dependency everywhere. Observables are the best solution, despite the limited support in WPF (need to use manual subscriptions).

#26