Support post-startup profile resets.
Reworks the new Profile model to store the Policy instance inside the Profile itself, which in turn allows us to do the same "Reset" operations supported in older versions (reverting to the "recommended" default or face plugin).

Includes a minor update to the `ProfileViewModel` to select a specific NPC - also included in the older code and needed for hotlinking from the Build page.

#26