Add component-based repository search.
This is important for the build, which - even though parts are being written to use resources found anywhere - still needs mod/component-specific searches for the facegen files.

Because the repositories are already optimized for global searches, the implementation doesn't try to further optimize for component-based searches. It is just a global search, filtered by component.

#26 #27