Mutable Mutagen env wraper.
Build tasks may be a little too complex to rewrite to be library-agnostic (for now), but in order to make them testable and work with IoC, they can still accept a limited interface with mutation support, rather than the entire `GameEnvironmentState`.

Moves the non-app-specific Mutagen extension methods into the wrapper library.

#26