Initial rewrite of the build system.
Intent is to have a loosely-coupled arrangement of tasks, with a pipeline system to handle all the scheduling and dependencies. This will allow us to take some of the extremely complex tasks (like the former `MutagenMergedPluginBuilder` and `MergedFolder`) and turn them into manageable units of work, and maybe improve parallelism.

Currently not tested at all, there are probably several bugs in here, but as yet none of this is linked to the main app. The initial graph does assemble with Autofac.

#26 #37 #64 #74