UI components for new build pipeline.
New system uses too many tasks to fit comfortably in the old UI. We now need a grid-based system to handle it.

Some hacks are necessary in order to deal with the space WPF insists on reserving for the scroll bar, and which we can't get rid of with a custom `ControlTemplate` due to ModernWpf's overrides.

#26