Full cutover to engine rewrites.
Replaces the analysis, profile and build components with their updated versions, and gets rid of most of the old code. Everything in the view model tree (i.e. other than WPF code-behind) is now wired with an Autofac container.

This has finally been tested with an end-to-end build, and although it is not at perfect parity, it is no longer practical to continue adding new code and trying to incrementally convert.

Critical tasks remaining are:
- Re-adding Vortex support
- Re-implementing the Facegen Dewiggifier as a new build task
- Making the loading screen usable again
- Show remaining/unstarted tasks during build
- Add help text for new build warnings and get rid of the old ones
- Show status screen at startup for unrestorable NPCs (might be better as long term feature)
- Tests for all the new code (although the old code had none either)
- May have missed a few other things that I hope will be uncovered in testing.

#26